Redraw health&energy bar with waving animation
This cost me 4-5 hours to find out the solution. There's no tutorial for this and I have to find the solution myself.  It's so difficult but I MADE IT!!

I drew a beautiful energy/health bar and waving aniamtion at the head of energy/health so it look like liquid now.